pyproject.toml
[tool.poetry]
name = "truelove"
version = "0.1.0"
description = ""
authors = ["ShiinaRinne <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
loguru = "^0.7.2"
requests = "^2.31.0"
asyncio = "^3.4.3"
aiohttp = "^3.9.1"
pydantic = "^2.5.3"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.24"
bilix = "^0.18.8"
aiosqlite = "^0.20.0"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
